Have notification bring user to search results

+++ QuickTap/app/src/main/java/com/example/android/quicktap/ResultsListActivity.java
@@ -31,7 +31,7 @@ public class ResultsListActivity extends AppCompatActivity {
         DBAssetHelper dbSetup = new DBAssetHelper(ResultsListActivity.this);
         dbSetup.getReadableDatabase();
 
-        mSearchId = getIntent().getIntExtra(QuickTapSQLiteOpenHelper.RESULTS_SEARCH_ID, -1);
+        mSearchId = (int) getIntent().getLongExtra(QuickTapSQLiteOpenHelper.RESULTS_SEARCH_ID, -1);
 
         //TODO - handle case where searchId = -1
         mHelper = new QuickTapSQLiteOpenHelper(ResultsListActivity.this);
